backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Caching
redis==5.0.1
hiredis==2.2.3

# Data Processing & ML
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.0

# Environment & Configuration
python-dotenv==1.0.0

# API & HTTP
httpx==0.25.2
requests==2.31.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing (optional, but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Demo & CLI
rich==13.7.0